Add module scripts
